# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /user/{user_id}:
    x-swagger-router-controller: user
    get:
      description: Get user profile by using user_id
      operationId: getUserProfileById
      parameters:
        - in: path
          name: user_id
          description: user_id of user to get profile
          required: true
          type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserProfileResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
...
definitions:
...
  MessageResponse:
    properties:
      message:
        type: string
...
  UserProfileResponse:
    required:
      - username
    properties:
      username:
        type: string
      myteam_list:
        type: array
        items:
          type: integer
      waitteam_list:
        type: array
        items:
          type: integer
...
securityDefinitions:
  api_key:
    type: apiKey
    in: query
    name: api_key
security:
  - api_key: []